Install packages and gems
Add the React front-end that will interact with the Rails API. Incorporate a starting set of packages
that I feel are likely to help any React-Rails application.

On Rails, this includes foreman, rack-cors, and rack-attack, with a number of other packages commented out.

On React, this includes Redux, Redux Thunk, superagent, and anything needed to connect those to React.